requirements.txt
# Core Dependencies
python-telegram-bot==20.6
requests==2.31.0
aiohttp==3.8.6
python-dotenv==1.0.0

# Data Validation and Settings
pydantic==2.4.2
pydantic-settings==2.0.3

# Web Framework (if needed)
fastapi==0.104.1
uvicorn==0.24.0
flask==2.3.3

# Database
asyncpg==0.28.0
psycopg2-binary==2.9.7

# Development Dependencies (install with -r requirements-test.txt for testing)